pyproject.toml
[project]
name = "FabricaNeeds-Back-Equipe5"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "", email = ""},
]
dependencies = [
    "django>=4.2.13",
    "djangorestframework>=3.15.1",
    "dotenv>=0.0.5",
    "python-dotenv>=1.0.1",
    "psycopg2>=2.9.9",
    "passage-identity>=2.5.0",
    "openapi-python-client>=0.21.1",
    "psycopg>=3.1.18",
    "psycopg2-binary>=2.9.9",
    "uvicorn>=0.29.0",
    "gunicorn>=22.0.0",
    "django-cors-headers>=4.3.1",
    "drf-spectacular>=0.27.2",
    "setuptools>=70.2.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"

[tool.pdm.scripts]
dev = "python manage.py runserver localhost:8000"
runserver = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"